requirements.txt
# Core dependencies
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0

# Enhanced article processing dependencies
beautifulsoup4>=4.12.0  # For web scraping features
feedparser>=6.0.0       # For RSS feed parsing if needed
trafilatura>=1.6.0      # For article text extraction
newspaper3k>=0.2.8      # Fallback for article extraction
tiktoken>=0.5.0         # For token counting
lxml_html_clean>=0.4.0  # Required for trafilatura/lxml compatibility

# Optional dependencies for audio processing
pydub>=0.25.0           # For combining audio chunks (optional)